git-svn-id: https://svn.win.tue.nl/repos/prom/Packages/InductiveVisualMiner/Trunk@19224 a68f48d4-8d00-424f-b812-3f22a534a02a

+++ src/org/processmining/plugins/inductiveVisualMiner/helperClasses/GraphvizDirectlyFollowsGraph.java
@@ -79,9 +79,9 @@ public class GraphvizDirectlyFollowsGraph {
 
 		//add the directly-follows edges
 		for (long edge : dfg.getDirectlyFollowsGraph().getEdges()) {
-			XEventClass from = dfg.getDirectlyFollowsGraph().getEdgeSource((int) edge);
-			XEventClass to = dfg.getDirectlyFollowsGraph().getEdgeTarget((int) edge);
-			int weight = (int) dfg.getDirectlyFollowsGraph().getEdgeWeight((int) edge);
+			XEventClass from = dfg.getDirectlyFollowsGraph().getEdgeSource(edge);
+			XEventClass to = dfg.getDirectlyFollowsGraph().getEdgeTarget(edge);
+			int weight = (int) dfg.getDirectlyFollowsGraph().getEdgeWeight(edge);
 
 			DotNode source = activityToNode.get(from);
 			DotNode target = activityToNode.get(to);
